Two new potential spells
tarot card : shot a little card in fornt of you. Low mana cost and very
small damage. The heal spell nemesis. Got a small chance that the card
launcher draw "the fool" (google is your friend), that in flict more
damage. It do a flaregun sound effect when the fool is draw (not sure if
it's a good idea)

cloud : summon a small cloud that stay for 7 seconds. It disappear,
then.

then : gonna work on some kind of ligthning bolt, and some kind of
poison.