Don't mind
sorry, I haven't used the base at all(probably because I raged and
started all from the very beginning).
the code is a mess right now, but I will make it more understandable
later.
it is very unstable, and please don't mind it, I think I have to come up
with a better idea of how it has to be done,
but for now, it's all I have =P